move align utils to misc_utils
